[move]: move notice view model to directory
