+++ lib/src/views/home/sections/notice_carousel_section.dart
@@ -5,7 +5,7 @@ import 'package:go_router/go_router.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 
 import '../../../utilities/design_system/app_texts.dart';
-import '../../../viewmodels/notice_view_model.dart';
+import '../../../viewmodels/notice/notice_view_model.dart';
 
 class NoticeCarouselSection extends StatefulWidget {
   NoticeCarouselSection({super.key});
